+++ src/services/identity/local.ts
@@ -1,23 +1,31 @@
+import * as dotenv from 'dotenv'
 import type {
-	IIdentifier, CredentialPayload, VerifiableCredential,
+	IIdentifier, 
+	CredentialPayload, 
+	VerifiableCredential,
 } from '@veramo/core'
 import { MemoryPrivateKeyStore } from '@veramo/key-manager'
 import { KeyManagementSystem } from '@veramo/kms-local'
 import { CheqdDIDProvider, ResourcePayload } from '@cheqd/did-provider-cheqd'
 import type {
-	BulkRevocationResult, BulkSuspensionResult, BulkUnsuspensionResult, CreateStatusList2021Result,
+	BulkRevocationResult,
+	BulkSuspensionResult,
+	BulkUnsuspensionResult,
+	CreateStatusList2021Result,
 } from '@cheqd/did-provider-cheqd/build/types/agent/ICheqd'
 import { CheqdNetwork } from '@cheqd/sdk'
-
 import {
-	BroadCastStatusListOptions, CreateStatusListOptions, CredentialRequest, DefaultRPCUrl,
-	StatusOptions, UpdateStatusListOptions,
+	BroadCastStatusListOptions,
+	CreateStatusListOptions,
+	CredentialRequest,
+	DefaultRPCUrl,
+	StatusOptions,
+	UpdateStatusListOptions,
 } from '../../types/shared.js'
 import { Connection } from '../../database/connection/connection.js'
 import { Veramo } from './agent.js'
-import { DefaultIdentity } from './default.js'
+import { DefaultIdentity } from './IIdentity.js'
 
-import * as dotenv from 'dotenv'
 dotenv.config()
 
 const {
