+++ src/services/identity/unauthorized.ts
@@ -1,13 +1,12 @@
+import * as dotenv from 'dotenv'
 import { MemoryPrivateKeyStore } from '@veramo/key-manager'
 import { KeyManagementSystem } from '@veramo/kms-local'
-
 import { DefaultRPCUrl } from '../../types/shared.js'
 import { Connection } from '../../database/connection/connection.js'
 import { Veramo } from './agent.js'
 import { CheqdDIDProvider } from "@cheqd/did-provider-cheqd";
 import { CheqdNetwork } from '@cheqd/sdk'
-import * as dotenv from 'dotenv'
-import { DefaultIdentity } from './default.js'
+import { DefaultIdentity } from './IIdentity.js'
 
 dotenv.config()
