+++ src/static/swagger.json
@@ -905,34 +905,18 @@
           },
           {
             "in": "query",
-            "name": "resourceName",
-            "description": "Filter a DID-Linked Resource query by Resource Name. See <a href=\"https://docs.cheqd.io/identity/credential-service/did-linked-resources/understanding-dlrs/technical-composition\">DID-Linked Resources</a> for more details.",
+            "name": "versionId",
+            "description": "Unique UUID version identifier of DID Document. Allows for fetching a specific version of the DID Document. See <a href=\"https://docs.cheqd.io/identity/architecture/adr-list/adr-001-cheqd-did-method#did-document-metadata\">cheqd DID Method Specification</a> for more details.",
             "schema": {
-              "type": "string"
+              "type": "string",
+              "format": "uuid"
             },
-            "example": "Cheqd Issuer Logo"
+            "example": "3ccde6ba-6ba5-56f2-9f4f-8825561a9860"
           },
           {
             "in": "query",
-            "name": "resourceType",
-            "description": "Filter a DID-Linked Resource query by Resource Type. See <a href=\"https://docs.cheqd.io/identity/credential-service/did-linked-resources/understanding-dlrs/technical-composition\">DID-Linked Resources</a> for more details.",
-            "schema": {
-              "type": "string"
-            },
-            "example": "IssuerLogo"
-          },
-          {
-            "in": "query",
-            "name": "resourceVersion",
-            "description": "Filter a DID-Linked Resource query by Resource Version, which is an optional free-text field used by issuers (e.g., \"v1\", \"Final Version\", \"1st January 1970\" etc). See <a href=\"https://docs.cheqd.io/identity/credential-service/did-linked-resources/understanding-dlrs/technical-composition\">DID-Linked Resources</a> for more details.",
-            "schema": {
-              "type": "string"
-            }
-          },
-          {
-            "in": "query",
-            "name": "resourceVersionTime",
-            "description": "Filter a DID-Linked Resource query which returns the closest version of the Resource *at* or *before* specified time. See <a href=\"https://docs.cheqd.io/identity/credential-service/did-linked-resources/understanding-dlrs/technical-composition\">DID-Linked Resources</a> for more details.",
+            "name": "versionTime",
+            "description": "Returns the closest version of the DID Document *at* or *before* specified time. See <a href=\"https://docs.cheqd.io/identity/architecture/adr-list/adr-005-did-resolution-and-did-url-dereferencing\">DID Resolution handling for `did:cheqd`</a> for more details.",
             "schema": {
               "type": "string",
               "format": "date-time"
@@ -941,20 +925,34 @@
           },
           {
             "in": "query",
-            "name": "resourceMetadata",
-            "description": "Return only metadata of DID-Linked Resource instead of actual DID-Linked Resource. Mutually exclusive with some of the other parameters.",
+            "name": "transformKeys",
+            "description": "This directive transforms the Verification Method key format from the version in the DID Document to the specified format chosen below.",
             "schema": {
-              "type": "boolean"
+              "type": "string",
+              "enum": [
+                "Ed25519VerificationKey2018",
+                "Ed25519VerificationKey2020",
+                "JsonWebKey2020"
+              ]
             }
           },
           {
             "in": "query",
-            "name": "checksum",
-            "description": "Request integrity check against a given DID-Linked Resource by providing a SHA-256 checksum hash. See <a href=\"https://docs.cheqd.io/identity/credential-service/did-linked-resources/understanding-dlrs/technical-composition\">DID-Linked Resources</a> for more details.",
+            "name": "service",
+            "description": "Query DID Document for a specific Service Endpoint by Service ID (e.g., `service-1` in `did:cheqd:mainnet:7bf81a20-633c-4cc7-bc4a-5a45801005e0#service-1`). This will typically redirect to the Service Endpoint based on <a href=\"https://w3c-ccg.github.io/did-resolution/#dereferencing\">DID Resolution specification</a> algorithm.",
             "schema": {
               "type": "string"
             },
-            "example": "dc64474d062ed750a66bad58cb609928de55ed0d81defd231a4a4bf97358e9ed"
+            "example": "service-1"
+          },
+          {
+            "in": "query",
+            "name": "relativeRef",
+            "description": "Relative reference is a query fragment appended to the Service Endpoint URL. **Must** be used along with the `service` query property above. See <a href=\"https://w3c-ccg.github.io/did-resolution/#dereferencing\">DID Resolution specification</a> algorithm for more details.",
+            "schema": {
+              "type": "string"
+            },
+            "example": "/path/to/resource"
           }
         ],
         "responses": {
